[TASK] Move all backend related ViewHelpers to Backend folder
This PR:

* Moves all backend related ViewHelpers to the backend folder
* Remove the ImplodeViewHelper (not needed anymore)

Fixes: #1313
