--- Classes/ViewHelpers/Format/ImplodeViewHelper.php
@@ -1,67 +1,0 @@
-<?php
-
-namespace ApacheSolrForTypo3\Solr\ViewHelpers\Format;
-
-/***************************************************************
- *  Copyright notice
- *
- *  (c) 2013-2015 Ingo Renner <ingo@typo3.org>
- *  All rights reserved
- *
- *  This script is part of the TYPO3 project. The TYPO3 project is
- *  free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2 of the License, or
- *  (at your option) any later version.
- *
- *  The GNU General Public License can be found at
- *  http://www.gnu.org/copyleft/gpl.html.
- *
- *  This script is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  This copyright notice MUST APPEAR in all copies of the script!
- ***************************************************************/
-
-use ApacheSolrForTypo3\Solr\ViewHelpers\AbstractSolrViewHelper;
-use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
-use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithRenderStatic;
-
-/**
- * ViewHelper to implode an array with a specific glue value.
- *
- * @author Timo Hund <timo.hund@dkd.de>
- */
-class ImplodeViewHelper extends AbstractSolrViewHelper
-{
-    use CompileWithRenderStatic;
-
-    // Set $escapeOutput to false in order to ensure that "newline" is handled correctly
-    protected $escapeOutput = false;
-
-    /**
-     * Initializes the arguments
-     */
-    public function initializeArguments()
-    {
-        parent::initializeArguments();
-        $this->registerArgument('glue', 'string', 'The glue value', false, ',');
-        $this->registerArgument('value', 'array', 'The the array to implode', false, []);
-    }
-
-    /**
-     * @param array $arguments
-     * @param \Closure $renderChildrenClosure
-     * @param RenderingContextInterface $renderingContext
-     * @return string
-     */
-    public static function renderStatic(array $arguments, \Closure $renderChildrenClosure, RenderingContextInterface $renderingContext)
-    {
-        $glue = isset($arguments['glue']) ? $arguments['glue'] : ',';
-        $value = isset($arguments['value']) ? $arguments['value'] : [];
-
-        return implode($glue, $value);
-    }
-}
